worksでgetStaticPropsを使用
